refactor: restructure letters for protected routes
