+++ src/app/(auth-protected)/letters/_components/filter-bar/favorite-filter.tsx
